included all chagnes and bugfixes from OTSDAQ running at FNAL - this should hopefully work and have all the fixes

+++ include/DTCSupervisor/SupervisorGUI.h
@@ -205,7 +205,7 @@ namespace Ph2TkDAQ {
         {
             bool bFound = false;
 
-            for (int iName = 0; iName < lstNames.size(); iName++)
+            for (size_t iName = 0; iName < lstNames.size(); iName++)
             {
                 if (!strImage.compare (lstNames[iName]) )
                 {
